Fix: Mobile: ICCM Diarrhoea Referral condition changes based on Number of days , Bloody diarrhoea and signs. 	- MotherNeonate Anc bug fixes and BP&Weight issuse fixes 	- MDT-14410

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/data/PregnancyDetailsModel.kt
@@ -3,7 +3,7 @@ package com.medtroniclabs.spice.data
 data class PregnancyDetailsModel(
     var height: Double? = null,
     var weight: Double? = null,
-    var pulse: Int? = null,
+    var pulse: Double? = null,
     var lastMenstrualPeriod: String? = null,
     var estimatedDeliveryDate: String? = null,
     var gestationalAge: Long? = null,
